+++ src/pages/Customers/CustomerDetail.tsx
@@ -151,7 +151,7 @@ export default function CustomerDetail() {
   }
 
   return (
-    <div className="max-w-4xl">
+    <div className="w-full">
       <div className="flex items-start justify-between mb-6">
         <div>
           <PageTitle>
@@ -215,6 +215,15 @@ export default function CustomerDetail() {
               <DescriptionDetails>
                 {customer.billing_address.country}
               </DescriptionDetails>
+
+              {customer.vat_id ? (
+                <>
+                  <DescriptionTerm>
+                    <Trans>VAT ID</Trans>
+                  </DescriptionTerm>
+                  <DescriptionDetails>{customer.vat_id}</DescriptionDetails>
+                </>
+              ) : null}
             </DescriptionList>
           </div>